Refactor api.py: extract Skills + Plugins + session-skill routes
Continues the api.py split. Stacks on top of #364.

- Skill catalog management (9 endpoints) → routes/skills.py
- Skill discovery (from-md, from-git, discover) → routes/skills.py
- Plugin management (5 endpoints) → routes/skills.py
- Session-scoped skill toggles (3 endpoints) → routes/skills.py

Total: 20 endpoints moved.

Deliberately left in api.py for now:
- /agents/{name}/skills/* (7 endpoints): they call the streaming
  helpers (_disconnect_streaming_sessions, _start_streaming_session,
  _get_streaming_restart_guard, _guard_message) which are deep
  closures over broker, manager, and per-agent state. Pulling those
  through ctx is its own meaningful change — better landed alongside
  the agents-domain extraction.

api.py: 9,179 → 8,816 lines (-363, -4.0%)
Cumulative since main: 11,138 → 8,816 (-2,322 lines, -20.8%)

Tests: 1542 passed, 1 skipped (federation tests need nacl, pre-existing).
